checks to make sure arguments being backed-up are regular files

+++ backup.sh
@@ -14,20 +14,23 @@ LAST_ARG=${!#};
 
 if [ $# -gt 0 ]; then  # if arguments exist
 
+    ARG_ARRAY=( $@ );
+    ARRAY_LENGTH=$#;
+
     if [ -d $LAST_ARG ] ; then # if the last argument is a directory..
 	BACKUP_DIRECTORY=$LAST_ARG; # then make it the backup directory
-	ARG_ARRAY=( $@ );
-	ARRAY_LENGTH=$#;
 	ARGS=${ARG_ARRAY[@]:0:$ARRAY_LENGTH-1}; # use all but last argument
 
     else # otherwise store files in designated backup folder
 	BACKUP_DIRECTORY=~/.backups;
-	ARGS=( $@ );
+	ARGS=${ARG_ARRAY[@]:0:$ARRAY_LENGTH}; # use all but last argument
     fi
 
     # loop through the arguments
     for i in $ARGS; do # loop through arguments
-	cp $i $i\_BAK_$NOW
-	mv $i\_BAK_$NOW $BACKUP_DIRECTORY;
+	if [ -f $i ]; then # if file is not a directory or anything special
+	    cp $i $i\_BAK_$NOW
+	    mv $i\_BAK_$NOW $BACKUP_DIRECTORY;
+	fi
     done
 fi
